src/requirements.txt
boto3
certifi
chardet
click
Django
django-advanced-filters
django-appconf
django-autocomplete-light
django-braces
django-crispy-forms
crispy-bootstrap3
django-easy-select2
django-extensions
django-haystack
django_compressor
django-querysetsequence
django-rq
Django-Select2
django-spurl
django-storages
django-taggit
elasticsearch==5.5.3
uvicorn
idna
model-mommy
psycopg2-binary
pytz
redis
requests
rq
urllib3
